requirements.txt
torch 
torchvision 
transformers 
requests 
pillow 
numpy
opencv-python 
tqdm
accelerate
ultralytics
numba
scikit-image
filterpy
timm
